+++ lib/xeroizer.rb
@@ -11,7 +11,7 @@ require 'nokogiri'
 require 'builder'
 require 'time'
 require 'bigdecimal'
-require 'cgi'
+require 'cgi/escape'
 require 'json'
 
 $: << File.expand_path(File.dirname(__FILE__))